updates to docs
