+++ pages/docs-framebuffer-hooks.js
@@ -1,0 +1,9 @@
+import React from 'react'
+import MarkdownElement from 'docs/MarkdownElement'
+import docs from 'packages/core/docs/framebuffers.md'
+
+function FramebufferHooks() {
+  return <MarkdownElement docs={docs} />
+}
+
+export default FramebufferHooks
